Add Blender source file.
This is a cleaner edition of the source file. The original required a
lot of manual tweaking after applying the Boolean operations. This one
can in theory be used as-is, although it still is recommended to remove
duplicate points and degenerate geometry, and manually fixing weird
topology after applying the operators.