Another attempt at clarifying the licensing.
If there's one thing I have learnt over the years, it is that if you
want to tell people that doing a certain thing is not necessary, and you
do not say this in exactly the perfect wording, then this imperfect text
will actually incite more people to do the unnecessary thing.
In this case, people keep asking me for explicit permission to sell this
model. You DO NOT need explicit permission, it's the whole point of the
CC-BY-SA license. I hope the new formulation makes this truly obvious.